docs(readme): updating instructions for use
docs(readme): updating instructions for use

Add type property for webpack to work correctly when converting files to webp format.

This addition helped me solve the problem of converting images to webp format when building a project. I hope this add-on will also help other developers when working with webpack.